use lookup table instead of searching for event
This will make the lookup function much faster at the cost of taking
more memory. And, will also pave a way to make easier changes in the
future. Also, fixed a bug of sending EV_ABS events with value 0 if the
two sides of the ABS are pressed at the same time (i.e dpad left and
right pressed together then left is released).