fix wrong neutral events for dpad
So, there is two event types for the dpad (EV_HAT0X and EV_HAT0Y)
And there is only one neutral state written in config.h (which is for
EV_HAT0X). Which causes the up and down buttons on the dpad to never
release. after this commit there will be two seperate events for the
neutral dpad (DPAD_XNEUTRAL and DPAD_YNEUTRAL) and will be handled in
the transilation accordingly.